Appoverviewwindow and dock bar
Tabs and Esc keys are not responding because key events are intercepted by the underlying layer. Improved the layout to resolve the issue of exiting when clicking the input box. Fixed the problem of the dock bar being too narrow at non-standard DPI. Added the functionality to switch to applications by pressing Tab after entering search terms on the list page.